chore(release): v1.6.1 — clear OpenClaw scanner warning, surface OGG/Opus in plugin info
CLI-only release. Engine v1.6.0 unchanged.

- Extract version-marker helpers to src/engine-version-marker.ts so
  src/engine-install.ts no longer co-imports readFileSync alongside
  fetch(). OpenClaw's plugin scanner pairs that into a
  [potential-exfiltration] warning at install time — false positive for
  the engine-binary download path. Re-exports preserve the public API
  for src/lib.ts, src/cli/install.ts, and the unit tests.
- openclaw.plugin.json: extend description with the kesha say
  --format ogg-opus capability so users see the TTS path in
  openclaw plugins inspect output.
- openclaw-plugin.cjs: top comment block documents kesha say usage
  and kesha install --tts prerequisite for messenger-ready voice notes.

Verified: bunx tsc --noEmit clean, full bun test suite passes
(110 unit + 28 integration), openclaw security audit --deep no longer
lists kesha-voice-kit, openclaw plugins doctor clean.
