+++ src/engine-install.ts
@@ -1,34 +1,21 @@
 import { dirname, join } from "path";
-import { existsSync, mkdirSync, chmodSync, readFileSync, writeFileSync } from "fs";
+import { existsSync, mkdirSync, chmodSync } from "fs";
 import { getEngineBinPath, getEngineCapabilities } from "./engine";
 import { log } from "./log";
 import { streamResponseToFile } from "./progress";
+import {
+  getVersionMarkerPath,
+  readInstalledEngineVersion,
+  writeInstalledEngineVersion,
+} from "./engine-version-marker";
+
+export {
+  getVersionMarkerPath,
+  readInstalledEngineVersion,
+  writeInstalledEngineVersion,
+} from "./engine-version-marker";
 
 const GITHUB_REPO = "drakulavich/kesha-voice-kit";
-
-/**
- * Version-marker file written next to the engine binary on download (#151).
- * Lets `kesha install` detect a stale cached engine after a CLI upgrade and
- * re-download without requiring `--no-cache`.
- */
-export function getVersionMarkerPath(binPath: string): string {
-  return `${binPath}.version`;
-}
-
-/** Read the recorded version for the installed binary, or null if missing / empty. */
-export function readInstalledEngineVersion(binPath: string): string | null {
-  try {
-    const v = readFileSync(getVersionMarkerPath(binPath), "utf-8").trim();
-    return v.length > 0 ? v : null;
-  } catch {
-    // Missing, unreadable, or permission-denied → treat as no marker so we re-download.
-    return null;
-  }
-}
-
-export function writeInstalledEngineVersion(binPath: string, version: string): void {
-  writeFileSync(getVersionMarkerPath(binPath), `${version}\n`);
-}
 
 function getEngineBinaryName(): string {
   const platform = process.platform;
